+++ src/common/string_util.h
@@ -1,21 +1,21 @@
-// Copyright (c) 2019-2022, The Monero Project
-// 
+// Copyright (c) 2026, The Shekyl Project
+//
 // All rights reserved.
-// 
+//
 // Redistribution and use in source and binary forms, with or without modification, are
 // permitted provided that the following conditions are met:
-// 
+//
 // 1. Redistributions of source code must retain the above copyright notice, this list of
 //    conditions and the following disclaimer.
-// 
+//
 // 2. Redistributions in binary form must reproduce the above copyright notice, this list
 //    of conditions and the following disclaimer in the documentation and/or other
 //    materials provided with the distribution.
-// 
+//
 // 3. Neither the name of the copyright holder nor the names of its contributors may be
 //    used to endorse or promote products derived from this software without specific
 //    prior written permission.
-// 
+//
 // THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS" AND ANY
 // EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED WARRANTIES OF
 // MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL
@@ -26,25 +26,55 @@
 // STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF
 // THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 
-// TODO(shekyl-v4): Replace boost::multiprecision with __uint128_t or a
-// lightweight big-integer library for the 128-bit division routines.
-#include <boost/multiprecision/cpp_int.hpp>
+#pragma once
 
-void div128_64(uint64_t dividend_hi, uint64_t dividend_lo, uint64_t divisor, uint64_t* quotient_hi, uint64_t *quotient_lo, uint64_t *remainder_hi, uint64_t *remainder_lo)
+#include <algorithm>
+#include <cctype>
+#include <string>
+#include <vector>
+
+namespace tools
 {
-  typedef boost::multiprecision::uint128_t uint128_t;
+namespace string_util
+{
+  inline void trim(std::string &s)
+  {
+    s.erase(s.begin(), std::find_if(s.begin(), s.end(),
+      [](unsigned char c) { return !std::isspace(c); }));
+    s.erase(std::find_if(s.rbegin(), s.rend(),
+      [](unsigned char c) { return !std::isspace(c); }).base(), s.end());
+  }
 
-  uint128_t dividend = dividend_hi;
-  dividend <<= 64;
-  dividend |= dividend_lo;
+  inline void trim_right(std::string &s, const std::string &chars)
+  {
+    s.erase(s.find_last_not_of(chars) + 1);
+  }
 
-  uint128_t q, r;
-  divide_qr(dividend, uint128_t(divisor), q, r);
+  inline void to_lower(std::string &s)
+  {
+    std::transform(s.begin(), s.end(), s.begin(),
+      [](unsigned char c) { return std::tolower(c); });
+  }
 
-  *quotient_hi = ((q >> 64) & 0xffffffffffffffffull).convert_to<uint64_t>();
-  *quotient_lo = (q & 0xffffffffffffffffull).convert_to<uint64_t>();
-  if (remainder_hi)
-    *remainder_hi = ((r >> 64) & 0xffffffffffffffffull).convert_to<uint64_t>();
-  if (remainder_lo)
-    *remainder_lo = (r & 0xffffffffffffffffull).convert_to<uint64_t>();
-}
+  inline bool iequals(const std::string &a, const std::string &b)
+  {
+    if (a.size() != b.size()) return false;
+    return std::equal(a.begin(), a.end(), b.begin(),
+      [](unsigned char ca, unsigned char cb) {
+        return std::tolower(ca) == std::tolower(cb);
+      });
+  }
+
+  inline std::string join(const std::vector<std::string> &v, const std::string &sep)
+  {
+    std::string result;
+    for (size_t i = 0; i < v.size(); ++i)
+    {
+      if (i > 0) result += sep;
+      result += v[i];
+    }
+    return result;
+  }
+
+} // namespace string_util
+} // namespace tools