+++ src/common/util.h
@@ -30,8 +30,7 @@
 
 #pragma once 
 
-#include <boost/thread/locks.hpp>
-#include <boost/thread/mutex.hpp>
+#include <mutex>
 #include <boost/optional.hpp>
 #include <system_error>
 #include <csignal>
@@ -217,8 +216,8 @@ namespace tools
     /*! \brief calles m_handler */
     static void handle_signal(int type)
     {
-      static boost::mutex m_mutex;
-      boost::unique_lock<boost::mutex> lock(m_mutex);
+      static std::mutex m_mutex;
+      std::unique_lock<std::mutex> lock(m_mutex);
       m_handler(type);
     }