+++ src/rpc/rpc_payment.h
@@ -32,7 +32,7 @@
 #include <unordered_set>
 #include <unordered_map>
 #include <map>
-#include <boost/thread/mutex.hpp>
+#include <mutex>
 #include <boost/serialization/version.hpp>
 #include "cryptonote_basic/blobdatatype.h"
 #include "cryptonote_basic/cryptonote_basic.h"
@@ -186,6 +186,6 @@ namespace cryptonote
     uint64_t m_nonces_stale;
     uint64_t m_nonces_bad;
     uint64_t m_nonces_dupe;
-    mutable boost::mutex mutex;
+    mutable std::mutex mutex;
   };
 }