+++ src/wallet/message_store.cpp
@@ -29,10 +29,9 @@
 
 #include "message_store.h"
 #include <boost/archive/portable_binary_iarchive.hpp>
-#include <boost/format.hpp>
-#include <boost/algorithm/string.hpp>
 #include <boost/system/error_code.hpp>
 #include <boost/filesystem.hpp>
+#include "common/string_util.h"
 #include <fstream>
 #include <sstream>
 #include "file_io_utils.h"
@@ -294,7 +293,7 @@ bool message_store::check_auto_config_token(const std::string &raw_token,
   {
     // Prefix must be there; accept it in any casing
     std::string raw_prefix(raw_token.substr(0, 3));
-    boost::algorithm::to_lower(raw_prefix);
+    tools::string_util::to_lower(raw_prefix);
     if (raw_prefix != prefix)
     {
       return false;
@@ -312,7 +311,7 @@ bool message_store::check_auto_config_token(const std::string &raw_token,
   }
 
   // Convert to strict lowercase and correct any common misspellings
-  boost::algorithm::to_lower(hex_digits);
+  tools::string_util::to_lower(hex_digits);
   std::replace(hex_digits.begin(), hex_digits.end(), 'o', '0');
   std::replace(hex_digits.begin(), hex_digits.end(), 'i', '1');
   std::replace(hex_digits.begin(), hex_digits.end(), 'l', '1');
@@ -585,8 +584,8 @@ size_t message_store::add_message(const multisig_wallet_state &state,
   // Save for every new message right away (at least while in beta)
   save(state);
 
-  MINFO(boost::format("Added %s message %s for signer %s of type %s")
-          % message_direction_to_string(direction) % m.id % signer_index % message_type_to_string(type));
+  MINFO("Added " << message_direction_to_string(direction) << " message " << m.id
+          << " for signer " << signer_index << " of type " << message_type_to_string(type));
   return m_messages.size() - 1;
 }
 
@@ -1105,8 +1104,9 @@ bool message_store::get_processable_messages(const multisig_wallet_state &state,
       else
       {
         // Don't sync, but give a hint how this minimal set COULD be synced if really wanted
-        wait_reason += (boost::format("\nUse \"mms next sync\" if you want to sync with just %s out of %s authorized signers and transact just with them")
-                                     % (m_num_required_signers - 1) % (m_num_authorized_signers - 1)).str();
+        wait_reason += "\nUse \"mms next sync\" if you want to sync with just "
+                       + std::to_string(m_num_required_signers - 1) + " out of "
+                       + std::to_string(m_num_authorized_signers - 1) + " authorized signers and transact just with them";
       }
     }
     if (sync)